Fix setup.py backwards compatibiity

+++ setup.py
@@ -82,10 +82,20 @@ class CMakeBuild(build_py):
         python_package_dir = os.path.join(cmake_build_dir,
                                           "tools", "torch-mlir", "python_packages",
                                           "torch_mlir")
-        shutil.copytree(python_package_dir,
-                        target_dir,
-                        symlinks=False,
-                        dirs_exist_ok=True)
+
+        if sys.version_info.minor >= 8:  # python >= 3.8
+            shutil.copytree(python_package_dir,
+                            target_dir,
+                            symlinks=False,
+                            dirs_exist_ok=True)
+
+        elif os.path.exists(python_package_dir):  # python <= 3.7
+            if os.path.exists(target_dir):
+                shutil.rmtree(target_dir, ignore_errors=False, onerror=None)
+
+            shutil.copytree(python_package_dir,
+                            target_dir,
+                            symlinks=False)
 
 
 class CMakeExtension(Extension):
